pyproject.toml
[project]
name = "genindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.21.1",
    "pandas>=2.3.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
    "tldextract>=5.3.0",
    "tqdm>=4.67.1",
]
